SW-1031 Run tests that depend on external services (#1176)
By default, we skip tests like `S3FileStoreTest` in CI because we don't want our
CI builds to fail if the AWS API is having a glitch that's unrelated to the
changes we're making to our code. But that means we don't detect problems in
the code under test, or, more commonly, problems caused by upgrades of the client
libraries we use to talk to those services.

Add a build step that runs those tests (which are now required to have class names
ending in `ExternalTest` but doesn't cause the whole workflow to fail if they
don't pass. The failures will still be visible in the build log, so we can decide
on a case-by-case basis whether we care about them.