SW-3697 Add "finished planting" flag to subzones
Allow admins to indicate that they have finished planting a planting subzone.

This doesn't do anything yet, but will affect some of the calculations we'll do
with observation data (e.g., estimating plant populations). Some of the
calculations depend on whether or not an entire planting zone is finished
planting; rather than tracking that as a user-modifiable flag, we'll instead
derive it from the subzones.