Track per-species observed plant totals (#1175)
We're going to be presenting various views of past observations that include
aggregated plant counts. Since a single monitoring plot can contain hundreds
of plants, and a single observation can contain dozens of monitoring plots,
doing the aggregation on the fly would be costly, especially for views that
show the results from many observations at once.

Add three tables to help reduce the aggregation cost. They record the number
of plants of each recording state (live, dead, existing), as well as total
plants and mortality rate (both of the latter excluding existing plants),
on a per-species, per-observation basis. Each table aggregates the results
at a different level of granularity: by monitoring plot, by planting zone,
and by planting site. They're updated (transactionally) when an observation
of a monitoring plot is completed.

For UI views that include per-species plant counts and mortality rates, we'll use
these new tables directly. For views that include total plant counts and mortality
rates across all species, we'll do on-the-fly aggregation using these tables; the
number of species in an observation will be small enough that this won't be
unacceptably costly.