Rename "Can't Tell" to "Unknown" (#1189)
Earlier versions of the product spec used the term "Can't Tell" to describe
species that the user can't identify at observation time, but now it's using the
term "Unknown." Update the data model and API to match the official terminology.

This is a breaking API change; clients that submit observation data will need to
replace `CantTell` with `Unknown` in the species certainty payload field.