EnumsTest shouldn't fail on extra values
If `Enums_en.properties` has a key that doesn't match any enum values, but does
have keys for all the values of all the enums, `EnumsTest` should pass.

This was causing CI failures after we deleted an enum value; the old English
display name was still in Phrase's string list and Phrase added it to back to
the properties file on export.