SW-3730 Drop total plants from observed species totals
The total number of observed plants can be calculated from the live, dead, and
existing totals.

The total was included as an explicit column in order to support exposing it
via the search API, but we've decided against using the search API to fetch
observation results, so it's no longer useful.