SW-3699 API endpoint to fetch observation photos
Add a GET endpoint to fetch photos of monitoring plots taken during observations.

The file IDs of the photos are already included in the observation results, so
there is no "list photos" endpoint. As with the other photo endpoints in the API,
optional `maxWidth` and `maxHeight` parameters can be included to retrieve
thumbnails.