SW-3722 Track subzone finished-planting time (#1193)
Some of the calculated observation results are conditional on whether or not all
the subzones in a zone or at a site are finished planting. The condition needs to
test whether the subzones were finished planting at the time of the observation,
not at the current time, or we'll calculate incorrect values for observations that
happened before a site was finished planting.

Replace the "finished planting" flag on subzones with a timestamp. The subzone API
payloads still include the boolean value, so this change should be transparent to
clients, except that the boolean field is no longer present in the search API.