SW-3717 Create ad-hoc planting sites from admin UI
Testing the geolocation features of the mobile app on real devices requires
visiting a planting site. For development testing by people working remotely,
it's too time-consuming to construct a real set of shapefiles for each person's
location, so people need to be able to create test planting sites near themselves.

Add a section to the organization page in the admin UI to allow creating a
planting site by drawing the boundary on a map. Sites created this way are very
simple (one zone and one subzone) but should be sufficient for test purposes.