SW-3731 Track plantings since last observation
Replace the `planting_site_populations` and `planting_subzone_populations` views
with tables that are updated when there are outplanting nursery withdrawals, and
add an analogous table for zone-level totals.

In addition to the all-time totals, these tables also have a new column that
tracks the number of plants added since the last observation. This number
is also adjusted when there are outplanting nursery withdrawals, but is reset to
zero when an observation is completed.

Tracking these numbers explicitly will reduce the cost of the queries to populate
the plants dashboard, especially once we add the "plants planted since last
observation" stat.

Once this code is deployed, a followup migration will backfill the values for
existing planting sites. The backfill needs to happen separately to avoid the case
where an outplanting withdrawal happens while this code is in the middle of being
deployed, and happens to be handled by a previous release which would fail to
update the numbers.