Add limit parameter to observation results API
To allow the plants dashboard to show statistics from the most recent
observation of a given planting site, allow the client to request a specific
number of results and change the result order so the most recently completed
observation always comes first. The client can thus pass in a planting site ID
and a limit of 1 to fetch the latest observation results for the site.

The default is still to return results from all observations.