Start migration subdirectories at multiples of 50 (#1198)
Make the database migration subdirectories start at multiples of 50 rather than at
multiples of 50 plus one. It's a little more intuitive if, e.g., migration 100 is
in the same directory as migration 101.

No changes to the contents of any migrations here, just renaming the directories
and moving the multiple-of-50-numbered migrations to their new homes.