Make observation results test more strict (#1203)
The tests in `ObservationResultsStoreTest` that compared calculated results
against a set of example spreadsheets was too lax and failed to catch some
errors in the spreadsheets. Rework them so that instead of verifying that each
spreadsheet row has a corresponding set of numbers in the calculated observation
results, they instead convert the observation results into a spreadsheet which
must be exactly equal to the spreadsheets with the expected values.

The previous version of the test failed to catch that we were returning non-live
species in the results, which violates the specs; fix that bug.

Freshen the spreadsheets with freshly-exported copies that fix a few problems the
original test didn't catch (a species row that shouldn't have been there, two
species rows that should have been there but weren't) and that express mortality
rates as rounded percentages since those are what we're returning in our
calculations.