SW-3348 Automatically log out incompatible sessions
When we make a non-backward-compatible change to one of the classes that is
stored as part of the user's login session data, treat the session as
nonexistent instead of bombing out with an internal error.

In practice, this will often be invisible to the user: they will be treated
as unauthenticated, so they'll be redirected to Keycloak. If Keycloak still
considers them logged in, they'll be issued a new access token and redirected
straight back to the web app which will then create a fresh session for them.

This is in preparation for moving away from the Keycloak adapter library; when
we switch away from it, sessions that include serialized objects from that
library will no longer be deserializable.