Increase heap size to fix Dokka failures
The recent upgrade to Dokka caused it to need more memory, which caused it to run
out of memory in CI builds. Bump up the maximum heap size to give it enough room
to run successfully, and move it a bit earlier in the CI build process so we'll
get faster failures if it breaks in the future.