SW-2710 Declare annotation property aliases
We have custom OpenAPI annotations to document that controller methods can
return certain kinds of responses. The annotations have parameters that supply
the values of some of the parameters of the underlying OpenAPI annotations.

Future versions of Spring are going to stop automatically using parameters on
annotations like these as the values of the underlying annotations, and will
require applications to explicitly declare that they want the parameters to be
treated as aliases for the parameters of other annotations. Spring Boot 3 logs
warnings about the upcoming behavior change.

Add the appropriate declarations so our annotations won't break in future Spring
versions.