Download GBIF data instead of uploading it
The admin UI allows importing GBIF species data from a zipfile. Previously, the
zipfile was uploaded by the browser using a file input. It is a huge file and
requires special treatment to avoid running into request size limits. And the
way file uploads are handled is different in Spring Boot 3, so we'd need to
implement a brand-new mechanism for dealing with the file.

Change the admin UI so it accepts a URL as input and pulls the file from there.
The common case will be in dev environments where the URL will point to the local
filesystem. It's not quite as convenient, but GBIF import is a pretty rare
operation so the added burden should be miniscule.