Stop supporting huge shapefiles (#1247)
Initially, we were going to include monitoring plot boundaries in uploaded
shapefiles for planting sites. Since there can be many thousands of monitoring
plots at a large site, this caused the shapefiles to be large enough that we
needed to stream their data in the application code rather than letting Spring
and Tomcat load them into memory or store them on disk.

But now we generate monitoring plot boundaries programmatically, and the
shapefiles are far smaller. So we can process them using the built-in file
upload support in Spring MVC. This is not only simpler, but avoids using an
API that's no longer supported in Spring Boot 3.