SW-2710 Remove Apache Commons FileUpload library (#1248)
We were using an Apache library to make it easier to work with uploads of large
files. Spring Boot 3 doesn't support that library; rather than implementing new
code to support large file uploads, we've instead removed the need to upload
large files in the first place. We can thus get rid of the library and rely on
Spring's built-in file upload support.