SW-3815 Read Google credentials from environment (#1252)
The credentials for the Google API, which we use to export report files to Google
Drive, have to be in JSON form. By default, the Google client library expects the
credentials to be in a file, but it also allows callers to directly supply a JSON
object.

Add a Terraware config value that lets us pass the JSON object in an environment
variable so we don't have to write it to a file on the server.