Allow enums to use string IDs
IUCN conservation categories are identified with two-letter codes. We'll want to
use those codes directly in our database schema rather than defining a synthetic
numeric ID like we do for our other enumerated types. We'll also want to use the
codes in the API, as opposed to our usual approach of using enum display names.

Update the enum code generator to support both integers and strings as enum IDs,
and to optionally allow the ID, rather than the name, to be used as the JSON
value.

Existing enums should continue to work as before.