SW-3127 Add IUCN conservation category (#1256)
Introduce IUCN conservation categories to our species data. The conservation
category will replace the "endangered" flag, but in this initial change, both
values are supported. If only one of the two is present when creating or
updating a species, the other is automatically populated.

A followup change will backfill the conservation categories for existing species.