SW-3683 Remove /api/v1/seedbank/values/all
The seedbank API has two endpoints to list all the values of a field. One of
them lists all the values that are currently present on any accessions, and
the other lists all possible values, even if they're not used.

The web and mobile apps no longer use the "list all possible values" endpoint,
so there's no need to maintain it. Removing it means we can remove a method
from `SearchTable` and its subclasses that only existed to support that endpoint.