Ignore irrelevant fields in plant observation payload
If a recorded plant's species is known, the server is supposed to ignore the
value of the `speciesName` field in the plot observation completion API
payload, but it doesn't; instead the request fails due to a database constraint
violation.

Update the code to conditionally ignore species ID and name like the API
documentation says it does.