Require planting sites to have at least 5 plots (#1283)
If you use the "draw site on map" feature of the admin UI to draw a planting site
that's only big enough for one 4-plot cluster, the server creates the planting
site but then can't start any observations because it can't allocate any temporary
plots.

Update the importer code to ensure that there are at least 5 plots available; if
not, site creation fails. Since plots are allocated in groups of 4, technically
"5 plots minimum" is really "8 plots minimum" and thus the planting site must be
at least 100x50 meters.