SW-3952 Reduce default confidence interval to 80%
Based on feedback from our carbon consultant, we only need an 80% confidence level
on observations, not 90%. Change the default Student's t value to 1.282, which
corresponds with an 80% confidence level, and reduce the default number of
permanent clusters and temporary plots based on the new value.

This only changes the default for new planting sites; existing ones can be edited
in the admin UI if needed.