SW-3977 Admin UI to create minimal site (#1317)
For testing observations in the mobile app, we want to be able to create the
smallest possible planting site, with one permanent cluster and one temporary
plot. Previously, this was impossible to do in the admin UI because the required
precision of the boundary lines was greater than the precision of the map.

Add a button to place a minimal-size planting site on the map. The user can move
it around on the map and change where the temporary plot is located relative to
the permanent cluster.

This requires adding some logic to the planting site creation code to account for
the fact that the map UI does coordinate transformations differently than the
server does, meaning that the site boundary submitted by the client will always
be off by a tiny amount from the server's point of view.