SW-4007 Use nursery time zone to check withdrawn date (#1320)
If a nursery withdrawal is dated earlier than the time a batch's remaining
quantity was most recently manually observed, we don't override the
manually-entered quantity.

Unfortunately, we were using UTC to figure out which date the last observation
fell on, meaning the system could erroneously think that a withdrawal happened
before the last observation instead of on the same day. This caused the system
to not update the quantities of batches when it should have.

A followup change will backfill the missing quantity updates.