SW-4041 Return permanent plot live/dead counts
The mortality rate graphs in the web app need to show live and dead plant totals
that only consider plants in permanent plots. We were already calculating those
totals on the server side since they are the inputs to the mortality rate
calculation; expose them in the observation results API payloads.