SW-4082 Calculate more observation plot counts
The mobile app needs to display a count of the number of monitoring plots in an
observation that haven't been completed yet, whether or not they've been claimed.
Add that count to the observation API, as well as a count of the total number of
monitoring plots in the observation regardless of status.