SW-3992 Add API endpoints for sub-locations
In preparation for adding sub-locations to nursery facilities, add a new set of
endpoints similar to the ones under `/api/v1/seedbank/storageLocations`. There's
no new functionality here, just a different API that isn't specific to seed banks.

In keeping with our current API structure conventions, the PUT and POST endpoints
expect facility IDs in URL paths instead of in request payloads.

The response payloads have a field for the number of active batches. This will be
populated in a later code change but is there from the start so as to reduce API
churn.