SW-4311 Implement monitoring plot replacement
Implement the business logic for replacing a monitoring plot in an observation.

The behavior varies depending on a combination of factors: whether the given
plot is permanent or temporary in the observation, whether or not there are
already completed plots at the site from the current or a prior observation,
whether the replacement request is temporary or long-term, and whether other
subzones are finished planting or not. Various permutations, as outlined in
the requirements document, are covered by test cases.

This feature introduces the concept of a monitoring plot existing at a site,
possibly having observation results from previous observations, but being
unavailable for inclusion in future observations. Add a flag to the monitoring
plots table to keep track of availability.