SW-4308 Add API for replacing monitoring plot
`POST /api/v1/tracking/observations/{observationId}/plots/{plotId}/replace` will
request that a monitoring plot in an observation be replaced.

Currently this triggers an email notification to be sent to the organization's
Terraware contact (or to the support email if there's no contact) but does not
actually cause the plot to be replaced. However, the API definition can be used
for client-side code generation.