Use Java 21 runtime
Previously, we were targeting Java 19 with the build (because that was the newest
version supported by the Kotlin compiler at one point) and using Java 20 as the
runtime JVM.

Update the runtime to use Java 21, and the build to target Java 20.

No functional change (we aren't using any of Java 21's new features) but it
should give us some minor performance improvements.