SW-4329 Assign short numbers to facilities
To support generating nursery batch numbers that include an indication of which
nursery the batch is in, add the concept of a "facility number." Facility numbers
are per-organization, per-facility-type. That is, for each organization, its first
nursery will have a facility number of 1, and its first seed bank will also have
facility number 1.