SW-4232 Return observation photo GPS information
Currently, when it uploads photos of a permanent monitoring plot's four corners,
the mobile app includes GPS coordinates and an indication of which corner the
photo shows. But this information was write-only and couldn't be retrieved for
later display.

Return it as part of the list of an observed plot's photos.