SW-4085 Rename storage_locations to sub_locations
Early versions of Terraware used the term "storage location" to describe fridges
and freezers in a seed bank. We renamed that concept to "sub-location" in the app,
but we continued to use the old name in the code.

In preparation for allowing users to add sublocations to nurseries, not just to
seed banks, start using the new name internally in the server.

This initial change is purely an internal refactoring. No changes to the API.