SW-4366 Move seed treatments to public schema
In preparation for allowing users to specify seed treatments for seedling batches,
move the enum table with the list of seed treatments out of the `seedbank` schema
and into `public` to indicate that it's not a seed-bank-specific resource, and
rename it from "viability test treatments" to "seed treatments" since it will be
used for things other than viability tests.

No client-visible behavior change here, just reorganization for clarity.