SW-4365 Carry accession ID across nursery transfers
If a batch was originally created via a nursery transfer from a seed bank, it will
have an accession ID indicating where it came from. If plants from the batch are
then transferred to a different nursery, the newly-created batch should have the
same accession ID so it can be traced back to its original seeds.