Record withdrawal IDs for incoming transfers
Previously, we only included the withdrawal ID on the batch quantity history
entries for the originating batch, but not for the destination batch; there
was thus no easy way to tell which withdrawal was responsible for a quantity
change.

Update the withdrawal code to include the withdrawal ID in the quantity
history on both sides of nursery transfers.