SW-4407 Fix adding user to org from admin UI (#1443)
Commit 7c9a28a introduced the ability for a super-admin to add a user to any
organization without the super-admin having to be a member of the organization.
Unfortunately, this would fail when the system tried to generate in-app and
email notifications to tell the user they'd been added.

For now, switch to the system user when generating these notifications. We will be
revamping the permission system for super-admins in the near future, at which
point we'll replace this with something more robust.