SW-4377 Search table for facility inventory totals
We're going to add a view to the web app that shows the total number of seedlings
across all species at each nursery. Currently, the inventory search tables all
group the totals by species ID. Add a new search table that ignores species ID and
aggregates all the batches at a facility. It includes a count of the total number
of species.

The new table can be searched using a payload like

    {
      "prefix": "facilities.facilityInventoryTotals",
      "fields": [
        "facility_id",
        "facility_name",
        "germinatingQuantity",
        "notReadyQuantity",
        "readyQuantity",
        "totalQuantity",
        "totalSpecies"
      ],
      "search": {
        "operation": "field",
        "field": "organization_id",
        "values": [
          "123"
        ]
      }
    }