SW-4367 Allow batches to have photos
Add endpoints to create, list, read, and delete photos of seedling batches.
The read endpoint supports thumbnails via the usual `maxWidth` and `maxHeight`
query string parameters.

The batch's history will include records of photos being created and deleted.
In the case of a deleted photo, the creation event doesn't include a file ID
since the file no longer exists in the system, but it still includes the user
ID and time.