SW-4368 Track changes to batch details
Add history tables to track changes to user-editable batch details including
sub-locations. Like the quantity history tale, each row in the details history
table is a snapshot of the values of a particular version.

Since some of the details are references to things that can be deleted
(projects, sub-locations), we also record the names of those things as of the
time the edit happened, so that we can show a sensible name when someone looks
at the batch's history afterwards.