Ignore no-op batch quantity edits
The web app has a single UI where both the non-quantity-related details and the
quantities can be edited, and when the user hits the save button, the app
updates both the details and the quantities even if the quantities didn't
actually change.

Currently, this results in a manual edit being recorded in the batch's quantity
history, which will show up as a line item in the history view once that's
implemented. Since the user didn't actually change the quantities, the line item
will be meaningless.

Add logic to detect and ignore no-op quantity edits.