SW-4439 Throw better error for unsupported image type
Currently, the content type validation in the image upload endpoints just looks at
the client-supplied MIME type and believes it. It's thus possible for a client to
say it's uploading a JPEG image but actually upload some other kind of file, and
the server will store it in the file store.

But when we then try to generate a thumbnail for that image, it will fail because
the Java ImageIO class doesn't know how to load images of unsupported types.

Previously, this resulted in a NullPointerException with no additional detail.
Add logic to try to detect the file type from its content and throw a more useful
exception so we can diagnose why thumbnail generation is failing.