SW-4429 Make "has no contact" a separate notification
Some notifications go to the organization's primary project contact, and if there
isn't a contact, they're instead sent to the support mailing list.

Previously, if we had to fall back on the support list, we would add a paragraph
to the text of the notification email about the organization not having a contact.
This meant that each individual notification would have to have conditional logic
to either include that paragraph or not.

Replace that approach with a separate "this org has no contact" notification.

Since the new notification is only ever sent to the support mailing list, its text
is included directly in the template files rather than coming from the localizable
strings list.