SW-4411 Rework planting site notification schema
We're about to add several more reminder notifications for planting sites, and
it will start to get a little unwieldy to have a separate column in the planting
sites table to keep track of each one.

Create a new table `planting_site_notifications` that tracks how many of each
type of notification we've sent for a planting site. This models the reminder
notifications a bit differently than before: they are considered to have the
same notification type but a different "notification number" where the first
notification about something is number 1, the first reminder is number 2, etc.