SW-4469 Show planting seasons in admin UI
List the planting seasons (including past ones) in the admin UI, and use the
planting season dates to prefill the default dates for the next observation.
Current and future planting seasons can be edited and new ones can be created,
but past ones can't be modified.

This replaces the planting season start/end month fields.