Upgrade to Spring Boot 3.2.0
Upgrade to the latest Spring Boot release and update some related dependencies
that support it.

* Flyway 10, which no longer bundles the PostgreSQL support, so we now include
  an additional library for that and the version number is now in
  `gradle.properties`.
* The jOOQ-Docker Gradle plugin, which now supports Flyway 10.
* JobRunr 6.3.3.

This version of Spring Boot includes a change to the format of serialized session
data, so update the example session in `SpringSessionCompatibilityTest`. The
instructions in that test's class docs were out of date, so update those too.