Use compound primary key for subzone populations
Commit c9ae259d replaced `plot_populations` with `planting_subzone_populations`.
There was an embedded type defined for `plot_populations` to represent its
compound primary key, but it wasn't replaced with one for the new table.

Update the embedded type and use it in the subzone populations search table;
the incorrect primary key on that table could have resulted in incorrect search
results, though luckily the web app isn't currently doing any searches that
would trigger the problem.