SW-4514 Support non-localizable enums
Some of our database-backed enums are only used internally, never shown directly
to users. Previously, we still had to add their display names to our strings table
(and those names would still be submitted for translation) because the generated
code would try to look them up.

Pull the localization functionality out of `EnumFromReferenceTable` into a new
interface `LocalizableEnum`. Enums that are exposed in the search API are required
to be localizable since they can potentially be exported in CSV files.

Whether or not an enum is localizable is controlled by an optional parameter to
`EnumTable` in the code generator config.