SW-4514 Remove EnumTable secondary constructors
In preparation for adding a flag to generated enums to control whether or not
their display names should be localizable, update the enum configuration to
use the primary constructor everywhere. The secondary constructors existed solely
to avoid having to use `listOf()` for single-use enums, but the minor cleanliness
gain will mostly go away once the call sites need a localization-related
parameter.