SW-4411 Detect when no planting season scheduled
Scan for planting sites that ought to have planting seasons but don't, and publish
application events to trigger an initial notification and then a reminder 4 weeks
after that.

The timing of the initial notification varies depending on whether or not the site
has had planting seasons in the past.

This change doesn't send any actual notifications, just publishes the application
events; the notifications will be implemented in a later change.