Fix jOOQ deprecation warning
We use a custom class for PostGIS geometry columns, and we have a converter that
plugs into jOOQ so it uses our class when reading and writing those columns.

When we were constructing a custom jOOQ `Field` object, we were telling jOOQ to
look up the converter for our class in its global list of converters. That pattern
is being deprecated per https://github.com/jOOQ/jOOQ/issues/15286 because it can
cause undefined behavior. The existing code still works, but causes jOOQ to log
warning messages.

Update our custom field construction to explicitly specify the converter it needs.