SW-4595 Fix incorrect apostrophes
ASCII apostrophe characters are special in properties files and have to be doubled
to escape them. But some text editors "helpfully" detect pairs of apostrophes and
turn the second one into a closing single quote mark. Then Phrase, or other
software, notices the remaining apostrophe and doubles it again, such that we end
up with strings that have an ASCII apostophe and a single quote together.

We also had a couple sentences that should have used the word "its" but instead
used "it’s."

Remove the unnecessary single quote characters and add a test to catch the problem
in the future.