SW-4573 Nursery species projects search table (#1596)
Add a new search table so clients can query which projects have active batches of
a particular species. This differs from the existing ability to query the project
for each batch in that a given project is only returned once even if it is
associated with multiple batches. Empty batches are not counted.

Example `POST /api/v1/search` request payload to return the list of project names
and IDs for species 123 along with its inventory totals:

    {
      "prefix": "inventories",
      "fields": [
        "germinatingQuantity",
        "notReadyQuantity",
        "readyQuantity",
        "totalQuantity",
        "projects.id"
        "projects.name"
      ],
      "search": {
        "operation": "field",
        "field": "species_id",
        "values": ["123"]
      }
    }