SW-4603 Allow accelerator admins to access admin UI (#1597)
We have an interceptor that restricts access to the admin UI based on an
annotation on the controller class. Previously, the interceptor only checked for
super-admins, but we are going to allow accelerator admins to access the UI as
well. Make the annotation take a list of global roles to allow.