Move AdminController to its own package
AdminController is getting big and it will be easier to maintain if it's split
into smaller classes. It's currently under the "customer" package which is
supposed to be for code related to managing users, organizations, and such, but
most of the admin UI's functionality is related to other things.

Move it to a new "admin" package.