Hardwire admin UI URL prefix
When the admin UI was first written, we wanted to allow for the possibility that
the frontend web server would do URL rewriting to expose it at a different path.
We thus needed to be able to easily change the path that was included in links
and form actions.

This has turned out not to be needed. Passing the prefix around and using it
everywhere adds noise to the templates and the controller classes.

Update AdminController to stop including the prefix in the models that get passed
to the templates, and update the templates to directly use `/admin` in their URL
paths. In many cases, this means attributes such as `href` can be string literals
instead of Thymeleaf expressions.