SW-4673 Delete plots when observation rescheduled (#1603)
If an observation is in progress but no plots are completed yet, we let admins
reschedule it for a future date. Get rid of the initial plot assignments so as
to calculate a fresh set of plots based on the state of the planting site on
the new start date.